Flush Log and Capture on Shared Lib Detach
On shared library detach we flush both open writeable files as cheap
insurance against a crash in another shared library during shutdown.

While this is unnecessary in the case of a clean shutdown as our dynamically
linked libc will flush and close files for us in its own detach
notification, a hard crash in another shared library such as a SEGV will
prevent the libc detach from ever happening.

Shutdown is a time in an application lifecycle during which it is easy for
a developer to ignore a crash that doesn't corrupt their data since it has
no negative effect on their users.

We would also like to be robust to use during development when
any of the shared libraries in the system may still be of beta or alpha
quality, and to be used alongside other layers that may be hacked together
and therefore of lower robustness as a deliberate design decision trading
development time versus utility.